Moved static vulkan objects to a container class
Cleaned up the app structure by moving some static (non-changing) vulkan objects out of the Application to a new RenderingBackend class. The idea is to later move creation of other components (e.g. swapchain) to methods on the backend in order to further abstract away the underlying Vulkan objects as working with them as raw handles everywhere is not very convenient.

Added ensureSuccess utility method to VkUtil as we had quite a few if result != VK_SUCCESS  lines throwing the same IllegalStateExceptions scattered about. An VkXXX API call (returning int result) is passed in as the first argument and the desired message as second. The method then performs the check and throws if necessary, concatenating the translated result to the message. The method call obfuscates the API call points a bit, but once I got used to it, I think its cleaner than throwing with explicit if-expressions everywhere.

Lastly, moved everything in Main to ApplicationRunner in order to make stuff non-static and to get semaphores inside an AutoCloseable wrapper. The main loop structure now feels much more intuitive when the cleanup/initialization is moved out of there.